Remove raf file
