--- src/client/raf.js
@@ -1,16 +1,0 @@
-const REQUEST_ANIMATION_FRAME = (() => {
-  return (
-    window.requestAnimationFrame ||
-    window.webkitRequestAnimationFrame ||
-    window.mozRequestAnimationFrame ||
-    window.oRequestAnimationFrame ||
-    window.msRequestAnimationFrame ||
-    function (callback) {
-      setTimeout(() => {
-        callback && callback();
-      }, 1000 / 60);
-    }
-  );
-})();
-
-export default REQUEST_ANIMATION_FRAME;
